refactor: separate transport, routing, and business logic in foreman
Extract four new controllers from the ~810-line ForemanWss:
- WorkerMessenger: outbound send infrastructure with admin broadcast
- ApiController: all /api/* REST route handlers from HttpServer
- WebhookController: webhook routing (routePrEvent, routeCheckEvent,
  routeIssueEvent, forwardEvent) registered via webhooks.onAny()
- ForemanWorkerController: all worker message business logic
  (hello handshakes, task assignment, reconnect, claim)

ForemanWss becomes a thin transport layer (WS setup, ping/pong,
upgrade, Map-based message router). HttpServer uses ApiController
and always passes a non-null Webhooks instance (dev mode uses
webhooks.receive() to bypass signature verification).

Closes #1064
