+++ CLAUDE.md
@@ -17,7 +17,7 @@ Two independent processes: **foreman** (server) and **worker** (agent). They com
 MVC structure. Models own state; controllers handle external inputs.
 
 - **Models** (`models/`) — `ActiveRecord` is the abstract base class. `Task`, `WebhookEvent`, `ForemanMessage`, `Repo`, `Worker`, and `Installation` are the main DB-backed models. `TaskManager` owns ephemeral in-memory state (branch mappings, blocker state) and encapsulates all issue/PR event lifecycle logic — one instance per repo, managed via a static registry. `Installation` represents a GitHub App installation; `repos.installation_id` is a nullable FK (null = App not in use). `Worker` tracks both in-memory (connected) and DB-persisted state; use the registry for live connections, `Worker.get()` for DB lookups. `activity-log.ts` provides a cross-table merge of webhook events and foreman messages for the admin dashboard — cross-table concerns like this don't belong to either model class.
-- **Controllers** (`controllers/`) — `http-server.ts` handles webhooks + REST + SPA. `wss.ts` is the WebSocket protocol layer for worker↔foreman communication; it delegates installation lifecycle webhooks to `InstallationsController`. `admin-ws.ts` broadcasts to the admin dashboard. `InstallationsController` handles GitHub App installation events (`installation.created/deleted`, `installation_repositories.added/removed`): creates/deletes `Installation` records, links/unlinks and activates/deactivates repos, and seeds tasks from GitHub.
+- **Controllers** (`controllers/`) — `http-server.ts` handles webhooks + REST + SPA (uses `ApiController` for API routes). `wss.ts` is a thin WebSocket transport layer (ping/pong, upgrade, connect/disconnect) that composes `ForemanWorkerController`, `WebhookController`, and `WorkerMessenger`. `ForemanWorkerController` owns all worker message business logic (hello handshakes, task assignment, reconnect, claim). `WebhookController` owns webhook routing (`routePrEvent`, `routeCheckEvent`, `routeIssueEvent`, `forwardEvent`) and registers via `webhooks.onAny()`. `WorkerMessenger` is the outbound send infrastructure (analogous to a renderer): all messages sent to workers go through it so it can log and broadcast to the admin dashboard. `admin-ws.ts` broadcasts to the admin dashboard. `InstallationsController` handles GitHub App installation events (`installation.created/deleted`, `installation_repositories.added/removed`): creates/deletes `Installation` records, links/unlinks and activates/deactivates repos, and seeds tasks from GitHub.
 - **Clients** (`clients/`) — `db-client.ts` wires the shared Supabase client. `GithubClient` supports two auth modes: personal-token and App installation token (auto-mints a short-lived token). Callers with an `Installation` record should use the installation auth mode.
 
 ### Agent/Worker (`src/agent/`)
