+++ src/foreman/controllers/http-server.ts
@@ -5,22 +5,19 @@ import { join, extname } from "path";
 import { fileURLToPath } from "url";
 import { Hono } from "hono";
 import { getRequestListener } from "@hono/node-server";
-import { Task } from "../models/task.js";
-import { Repo } from "../models/repo.js";
-import { Worker } from "../models/worker.js";
-import { queryActivityLog } from "../models/activity-log.js";
-import type { TaskStatus } from "../../../shared/wire.js";
+import { ApiController } from "./api-controller.js";
 import { fmtError, log } from "../../utils.js";
 
 export interface HttpServerOptions {
-  webhooks: InstanceType<typeof Webhooks> | null;
-  routeEvent: (id: string, name: string, payload: unknown) => void | Promise<void>;
+  webhooks: InstanceType<typeof Webhooks>;
+  /** When true, incoming webhook requests must carry a valid signature. */
+  verifySignature: boolean;
 }
 
 export class HttpServer {
   readonly server: http.Server;
 
-  constructor({ webhooks, routeEvent }: HttpServerOptions) {
+  constructor({ webhooks, verifySignature }: HttpServerOptions) {
     const app = new Hono();
 
   // ── Webhook ────────────────────────────────────────────────────────────────
@@ -35,7 +32,7 @@ export class HttpServer {
     }
 
     try {
-      if (webhooks) {
+      if (verifySignature) {
         if (!signature) {
           return c.text("Missing signature", 401);
         }
@@ -46,8 +43,13 @@ export class HttpServer {
           payload: rawBody,
         });
       } else {
-        const parsed = JSON.parse(rawBody) as unknown;
-        await routeEvent(id, name, parsed);
+        // receive() expects a parsed payload object (not raw string); the name
+        // string is dynamically typed at runtime so we cast the function.
+        await (webhooks.receive as (event: { id: string; name: string; payload: unknown }) => Promise<void>)({
+          id,
+          name,
+          payload: JSON.parse(rawBody),
+        });
       }
       return c.text("OK", 200);
     } catch (err) {
@@ -62,112 +64,7 @@ export class HttpServer {
   );
 
   // ── REST API ───────────────────────────────────────────────────────────────
-  app.get("/api/log", async (c) => {
-    try {
-      const before = c.req.query("before");
-      const entries = await queryActivityLog({ limit: 50, ...(before ? { before } : {}) });
-      return c.json(entries);
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/tasks/:id", async (c) => {
-    try {
-      const task = await Task.get(c.req.param("id"));
-      if (!task) return c.json({ error: "not found" }, 404);
-      return c.json(task.toWire());
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/tasks/:id/events", async (c) => {
-    try {
-      const taskId = c.req.param("id");
-      const before = c.req.query("before");
-      const entries = await queryActivityLog({ taskId, limit: 50, ...(before ? { before } : {}) });
-      return c.json(entries);
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/workers/:id", async (c) => {
-    try {
-      const workerId = c.req.param("id");
-      const worker = await Worker.get(workerId);
-      if (!worker) return c.json({ error: "not found" }, 404);
-      return c.json(worker.toWire());
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/workers/:id/messages", async (c) => {
-    try {
-      const workerId = c.req.param("id");
-      const before = c.req.query("before");
-      const entries = await queryActivityLog({ workerId, limit: 50, ...(before ? { before } : {}) });
-      return c.json(entries);
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/tasks", async (c) => {
-    try {
-      const statusFilter = c.req.query("status") as TaskStatus | undefined;
-      const tasks = await Task.list();
-      const filtered = statusFilter
-        ? tasks.filter((t) => t.status === statusFilter)
-        : tasks.filter((t) => t.status !== "complete");
-      return c.json(filtered.map((t) => t.toWire()));
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/repos", async (c) => {
-    try {
-      const repos = await Repo.list();
-      return c.json(repos.map((r) => r.toWire()));
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/repos/:owner/:repo", async (c) => {
-    try {
-      const fullName = `${c.req.param("owner")}/${c.req.param("repo")}`;
-      const repo = await Repo.findByFullName(fullName);
-      if (!repo) return c.json({ error: "not found" }, 404);
-      const installation = await repo.installation;
-      return c.json({ ...repo.toWire(), installation: installation?.toWire() ?? null });
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
-
-  app.get("/api/repos/:id/log", async (c) => {
-    try {
-      const repoId = Number(c.req.param("id"));
-      const before = c.req.query("before");
-      const entries = await queryActivityLog({ repoId, limit: 50, ...(before ? { before } : {}) });
-      return c.json(entries);
-    } catch (err) {
-      log(`ERROR API query failed: ${fmtError(err)}`);
-      return c.json({ error: "internal error" }, 500);
-    }
-  });
+  new ApiController().register(app);
 
   // ── Static files (React SPA) ───────────────────────────────────────────────
   app.use("*", async (c) => {
